refactoring code for language translation modal - 60% complete
