+++ src/app/components/language/lib/constants.ts
@@ -1,0 +1,5 @@
+export const UNSAVED_TRANSLATION_LOCAL_STORAGE = "unsavedTranslationState";
+export const LOCAL_STORAGE_DEBOUNCE_TIME = 2000
+export const SYNC_DEBOUNCE_TIME = 2000
+export const DEFAULT_LANGUAGE = "Español";
+
