+++ src/app/components/language/mainLangSelect.tsx
@@ -1,25 +1,22 @@
 "use client";
-const defaultLanguage = "Español";
 import React from "react";
-import type { Metadata } from "next";
-import compStyles from "@/components/components.module.css";
 import LanguageSelectorUser from "@/components/language/languageSelectorUser";
 import languageStyles from "@/components/language/language.module.css";
 import LanguageEditTransalte from "@/components/language/languageEditTranslate";
 import { useLanguageText } from "@/stores/languageLoad";
+import { DEFAULT_LANGUAGE } from "@/components/language/lib/constants";
 
 export default function MainLangSelect() {
-  const { getText, availableLanguages, setLanguage, selectedLanguage } =
-    useLanguageText();
+  const { setLanguage } = useLanguageText();
 
   React.useEffect(() => {
-    setLanguage(defaultLanguage);
+    setLanguage(DEFAULT_LANGUAGE);
   }, [setLanguage]);
 
   return (
     <div style={{ display: "flex", flexDirection: "row", gap: "0.2rem" }}>
       <LanguageSelectorUser
-        defaultLanguage={defaultLanguage}
+        defaultLanguage={DEFAULT_LANGUAGE}
         selectorClassName={`${languageStyles.language_selector} ${languageStyles.language_selector_navbar}`}
       />
       <LanguageEditTransalte />
