+++ src/app/components/language/types.d.ts
@@ -1,3 +1,6 @@
 export type LanguageText = {
     [key: string] : string
 }
+export type LanguageUpdateValues = {
+    [key: string] : boolean
+}
